pyproject.toml
[project]
name = "guipylib"
version = "0.0.0"
description = "UI library with GPU-accelerated rendering"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Casey Culbertson" },
    { name = "Jason Zhang", email = "me@jasonzhang.dev" },
]
requires-python = ">=3.7"
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
dependencies = [
    "moderngl>=5.8.0",
    "glfw>=2.6.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
]

[project.urls]
Homepage = "https://github.com/Zjjc123/guipy"
Repository = "https://github.com/Zjjc123/guipy"

[dependency-groups]
dev = [
    "Sphinx>=5.1.1",
    "sphinx-rtd-theme>=1.0.0",
    "sphinx-autoapi>=1.9.0",
    "black>=22.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["guipy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
